Remove stuff out of the runtime that does not belong there.
There might be more, but it is a start.
