+++ node/overseer/examples/minimal-example.rs
@@ -26,7 +26,8 @@ use futures::{
 };
 use futures_timer::Delay;
 
-use polkadot_primitives::v1::{BlockData, PoV};
+use polkadot_node_primitives::PoV;
+use polkadot_primitives::v1::BlockData;
 use polkadot_overseer::{Overseer, AllSubsystems};
 
 use polkadot_subsystem::{Subsystem, SubsystemContext, SpawnedSubsystem, FromOverseer};
